Hotel Management Software
this software can be used to maintain database of any hotel i.e. to stor the information of  guest as well as staff details
This software is in java language and an offline software which you can esily edit or open using netbeans 8.0 or above an also have a database file which is just have to be restore using MySQL software.